requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
yfinance==0.2.28
pandas==1.5.3
numpy==1.24.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
schedule==1.2.0
lxml==4.9.3
